+++ CHANGELOG.rst
@@ -4,6 +4,30 @@ servicenow.itsm Release Notes
 
 .. contents:: Topics
 
+v2.12.1
+=======
+
+Bugfixes
+--------
+
+- api - Remove six imports since py2 support is not needed
+- client, validation - Remove six imports since py2 support is not needed
+- eda/plugins/event_source/records - Add meta file so EDA plugin docs will be parsed by Galaxy (fixes https://github.com/ansible-collections/servicenow.itsm/issues/498)
+- records - Prevent missed events between polls by avoiding future 'since' timestamps; advance the 'since' timestamp (cursor) to the latest seen value at second precision (monotonic). (issue
+
+v2.12.0
+=======
+
+Release Summary
+---------------
+
+Add Event-Driven Ansible record polling feature and fix some CI/CD issues.
+
+Minor Changes
+-------------
+
+- records - add EDA event source plugin that allows you to monitor a table for new or updated records
+
 v2.11.0
 =======